eliminacion de testSites, modificacion de ubicacion de archivos
