--- D&Cv2/testSite.cpp
@@ -1,12 +1,0 @@
-#include <iostream>
-#include <string>
-#include <vector>
-using namespace std;
-int main(int argc, char const *argv[])
-{
-    
-    
-    
-    
-    return 0;
-}
